feat(pyd): move tqdm-progress consumer to pyd.consumers
